+++ simple-subagent/index.ts
@@ -1,7 +1,8 @@
 import extension from './extensions/extension.ts'
 
-export { renderLiveCompact } from './extensions/display.ts'
+export { renderSubagentWidget } from './extensions/display.ts'
 export { startJob } from './extensions/execute-subagents.ts'
+export { JobRegistry } from './extensions/jobs.ts'
 export type {
   ForkOverride,
   SubagentRequest,